+++ src/core/index.ts
@@ -4,6 +4,7 @@ export * from "./vec3.js";
 export * from "./vec4.js";
 export * from "./mat3.js";
 export * from "./quat.js";
+export * from "./transform3.js";
 export * from "./euler.js";
 export * from "./dual-quat.js";
 export * from "./policy.js";